Move scheduler-related errors and structs to their own module
I had moved them to `model` but that should be for the world state
stuff.

This adds a `tasks` module which exports various scheduler / task
interface related things like errors, TaskId, events
