+++ crates/console-host/src/main.rs
@@ -246,7 +246,7 @@ fn console_loop(
                     printer
                         .print(
                             (match msg.event() {
-                                moor_values::model::Event::TextNotify(s) => s,
+                                moor_values::tasks::Event::TextNotify(s) => s,
                             })
                             .to_string(),
                         )
