+++ crates/daemon/src/rpc_server.rs
@@ -36,8 +36,8 @@ use zmq::{Socket, SocketType};
 use moor_kernel::tasks::sessions::SessionError::DeliveryError;
 use moor_kernel::tasks::sessions::{Session, SessionError, SessionFactory};
 use moor_kernel::tasks::TaskHandle;
-use moor_values::model::SchedulerError::CommandExecutionError;
-use moor_values::model::{CommandError, NarrativeEvent, TaskId, TaskResult};
+use moor_values::tasks::SchedulerError::CommandExecutionError;
+use moor_values::tasks::{CommandError, NarrativeEvent, TaskId, TaskResult};
 use moor_values::util::parse_into_words;
 use moor_values::var::Objid;
 use moor_values::var::Symbol;
