+++ crates/kernel/benches/vm_benches.rs
@@ -28,11 +28,11 @@ use moor_kernel::tasks::sessions::{NoopClientSession, Session};
 use moor_kernel::tasks::task_scheduler_client::TaskSchedulerClient;
 use moor_kernel::tasks::vm_host::{VMHostResponse, VmHost};
 use moor_kernel::tasks::VerbCall;
-use moor_values::model::AbortLimitReason;
 use moor_values::model::CommitResult;
 use moor_values::model::VerbArgsSpec;
 use moor_values::model::{BinaryType, VerbFlag};
 use moor_values::model::{WorldState, WorldStateSource};
+use moor_values::tasks::AbortLimitReason;
 use moor_values::util::BitEnum;
 use moor_values::var::{List, Symbol};
 use moor_values::{AsByteBuffer, NOTHING, SYSTEM_OBJECT};
