+++ crates/kernel/src/tasks/scheduler_client.rs
@@ -25,8 +25,8 @@ use moor_values::var::{Objid, Var};
 
 use crate::tasks::sessions::Session;
 use crate::tasks::TaskHandle;
-use moor_values::model::SchedulerError;
-use moor_values::model::SchedulerError::CompilationError;
+use moor_values::tasks::SchedulerError;
+use moor_values::tasks::SchedulerError::CompilationError;
 
 /// A handle for talking to the scheduler from the outside world.
 /// This is not meant to be used by running tasks, but by the rpc daemon, tests, etc.
