+++ crates/kernel/src/tasks/suspension.rs
@@ -28,7 +28,7 @@ use moor_values::var::Objid;
 use crate::tasks::sessions::{NoopClientSession, Session, SessionFactory};
 use crate::tasks::task::Task;
 use crate::tasks::{TaskDescription, TasksDb};
-use moor_values::model::{TaskId, TaskResult};
+use moor_values::tasks::{TaskId, TaskResult};
 
 /// State a suspended task sits in inside the `suspended` side of the task queue.
 /// When tasks are not running they are moved into these.
