+++ crates/kernel/src/tasks/task.rs
@@ -35,10 +35,10 @@ use crossbeam_channel::Sender;
 use lazy_static::lazy_static;
 use tracing::{error, trace, warn};
 
-use moor_values::model::CommandError::PermissionDenied;
-use moor_values::model::WorldState;
-use moor_values::model::{CommandError, CommitResult, WorldStateError};
-use moor_values::model::{TaskId, VerbInfo};
+use moor_values::model::{CommitResult, VerbInfo, WorldState, WorldStateError};
+use moor_values::tasks::CommandError;
+use moor_values::tasks::CommandError::PermissionDenied;
+use moor_values::tasks::TaskId;
 use moor_values::util::parse_into_words;
 use moor_values::var::Symbol;
 use moor_values::var::{v_int, v_str};
@@ -467,7 +467,7 @@ impl Task {
             Err(WorldStateError::VerbPermissionDenied)
             | Err(WorldStateError::ObjectPermissionDenied)
             | Err(WorldStateError::PropertyPermissionDenied) => {
-                return Err(CommandError::PermissionDenied);
+                return Err(PermissionDenied);
             }
             Err(wse) => {
                 return Err(CommandError::DatabaseError(wse));
@@ -644,9 +644,9 @@ mod tests {
     use moor_compiler::{compile, Program};
     use moor_db_wiredtiger::WiredTigerDB;
     use moor_values::model::{
-        ArgSpec, BinaryType, CommandError, Event, PrepSpec, TaskId, VerbArgsSpec, VerbFlag,
-        WorldState, WorldStateSource,
+        ArgSpec, BinaryType, PrepSpec, VerbArgsSpec, VerbFlag, WorldState, WorldStateSource,
     };
+    use moor_values::tasks::{CommandError, Event, TaskId};
     use moor_values::util::BitEnum;
     use moor_values::var::Error::E_DIV;
     use moor_values::var::Symbol;
