+++ crates/kernel/src/tasks/task_scheduler_client.rs
@@ -16,8 +16,10 @@ use std::time::Instant;
 
 use crossbeam_channel::Sender;
 
-use moor_values::model::{CommandError, NarrativeEvent};
-use moor_values::model::{Perms, TaskId};
+use moor_values::model::Perms;
+use moor_values::tasks::{
+    AbortLimitReason, CommandError, NarrativeEvent, TaskId, UncaughtException,
+};
 use moor_values::var::Objid;
 use moor_values::var::Symbol;
 use moor_values::var::Var;
@@ -25,8 +27,6 @@ use moor_values::var::Var;
 use crate::tasks::task::Task;
 use crate::tasks::TaskDescription;
 use crate::vm::Fork;
-use moor_values::model::AbortLimitReason;
-use moor_values::model::UncaughtException;
 
 /// A handle for talking to the scheduler from within a task.
 #[derive(Clone)]
