+++ crates/values/src/tasks/mod.rs
@@ -12,15 +12,13 @@
 // this program. If not, see <https://www.gnu.org/licenses/>.
 //
 
-pub use crate::tasks::scheduler_client::SchedulerClient;
-pub use crate::tasks::suspension::{SuspendedTask, WakeCondition};
-pub use crate::tasks::task::Task;
-pub use crate::tasks::ServerOptions;
-pub use moor_values::tasks::TaskId;
+mod errors;
+mod events;
 
-pub mod builtins;
-pub mod config;
-pub mod matching;
-pub mod tasks;
-pub mod textdump;
-pub mod vm;
+pub use errors::{
+    AbortLimitReason, CommandError, SchedulerError, TaskResult, UncaughtException, VerbProgramError,
+};
+
+pub use events::{Event, NarrativeEvent};
+
+pub type TaskId = usize;
